updated product card and footer
I have updated the product card to ensure that all photos are the same size, regardless of the device they are viewed on. I have also added functionality to the footer, removing unnecessary blocks and adding a brand transfer option. To achieve this, I used the use Effect hook.
I apologize for using this method instead of hooks, as I did not fully understand the code structure at the time. However, together with my senior developer, we will work to fix this in the future.